Added multiple functionalities as described in commits

+++ app/src/main/java/com/omada/junctionadmin/data/repository/PostDataRepository.java
@@ -3,9 +3,13 @@ package com.omada.junctionadmin.data.repository;
 import android.net.Uri;
 import android.util.Log;
 
+import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
+import com.google.android.gms.tasks.Continuation;
+import com.google.android.gms.tasks.Task;
+import com.google.android.gms.tasks.Tasks;
 import com.google.common.collect.ImmutableList;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.DocumentReference;
@@ -13,6 +17,7 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
+import com.google.firebase.firestore.SetOptions;
 import com.google.firebase.firestore.WriteBatch;
 import com.google.firebase.storage.FirebaseStorage;
 import com.omada.junctionadmin.data.BaseDataHandler;
@@ -383,7 +388,9 @@ public class PostDataRepository extends BaseDataHandler {
                 .collection("posts")
                 .document(eventModel.getId());
 
-        batch.delete(eventDocRef);
+        Map<String, String> dataMap = new HashMap<>();
+        dataMap.put("status", "deleted");
+        batch.set(eventDocRef, dataMap);
 
         MainDataRepository.getInstance().getVenueDataRepository()
                 .deleteBooking(eventModel, batch);
@@ -391,28 +398,42 @@ public class PostDataRepository extends BaseDataHandler {
         batch.commit()
                 .addOnSuccessListener(aVoid -> {
 
-                    FirebaseStorage
-                            .getInstance()
-                            .getReference()
-                            .child("organizationFiles")
-                            .child(eventModel.getCreator())
-                            .child("posts")
-                            .child(eventModel.getId())
-                            .delete()
-                            .addOnCompleteListener(aVoid2 -> {
-                               Log.e("Posts", "Post image deletion successful");
-                            });
+                    /*
+                     The image has not been deleted because it needs to appear in user
+                     registered events
+                    */
 
                     MainDataRepository
                             .getInstance()
                             .getUserDataRepository()
                             .decrementHeldEventsNumber();
-
-                    resultLiveData.setValue(new LiveEvent<>(true));
                 })
                 .addOnFailureListener(e -> {
                     resultLiveData.setValue(new LiveEvent<>(false));
-                    Log.e("Post", "Error creating post");
+                    Log.e("Post", "Error deleting post");
+                })
+                .continueWithTask(task -> {
+                    if(!task.isSuccessful()){
+                        return Tasks.forException(new InterruptedException());
+                    }
+                    Log.e("Post", eventModel.getImage());
+                    return FirebaseStorage.getInstance()
+                            .getReference(eventModel.getImage().replaceFirst("gs://[A-Za-z0-9.-]*/", ""))
+                            .delete();
+                })
+                .addOnCompleteListener(task -> {
+                    if(task.isSuccessful()) {
+                        resultLiveData.setValue(new LiveEvent<>(true));
+                    }
+                    else if(task.getException() instanceof InterruptedException){
+                        Log.e("Post", "Post image not deleted because post deletion unsuccessful");
+                        task.getException().printStackTrace();
+                    }
+                    else {
+                        Log.e("Post", "Error deleting post image");
+                        task.getException().printStackTrace();
+                        resultLiveData.setValue(new LiveEvent<>(true));
+                    }
                 });
 
         return resultLiveData;
